requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
google-genai>=1.0,<2
pydantic>=2,<3
